refactor: Chat 패키지 분리
