+++ src/main/java/com/example/travelday/domain/chat/controller/ChatController.java
@@ -1,8 +1,8 @@
-package com.example.travelday.domain.travelroom.controller;
+package com.example.travelday.domain.chat.controller;
 
-import com.example.travelday.domain.travelroom.dto.request.ChatReqDto;
-import com.example.travelday.domain.travelroom.dto.response.ChatResDto;
-import com.example.travelday.domain.travelroom.service.ChatService;
+import com.example.travelday.domain.chat.dto.request.ChatReqDto;
+import com.example.travelday.domain.chat.dto.response.ChatResDto;
+import com.example.travelday.domain.chat.service.ChatService;
 import com.example.travelday.global.common.ApiResponseEntity;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -32,7 +32,7 @@ public class ChatController {
      * 채팅 메시지 전송
      */
     @MessageMapping("/{travelRoomId}/chat/send")
-    @SendTo("/topic/{travelRoomId}/chat")
+    @SendTo("/{travelRoomId}/chat")
     public ChatResDto sendChatMessage(@DestinationVariable("travelRoomId") Long travelRoomId, @Payload ChatReqDto chatReqDto, @AuthenticationPrincipal UserDetails userDetails) {
         return chatService.saveChat(travelRoomId, chatReqDto, userDetails.getUsername());
     }
