+++ src/main/java/com/example/travelday/domain/chat/dto/response/ChatMemberResDto.java
@@ -1,4 +1,4 @@
-package com.example.travelday.domain.auth.dto.response;
+package com.example.travelday.domain.chat.dto.response;
 
 import com.example.travelday.domain.auth.entity.Member;
 import com.fasterxml.jackson.annotation.JsonFormat;
